Module bcm2711_pac::dmac::DMA4_CS

Modules

ABORT

Abort DMA (W1SC)

ACTIVE

Activate the DMA4

DISDEBUG

Disable debug pause signal

DMA_BUSY

Indicates the DMA4 is busy (RO)

DREQ

DREQ state (RO)

DREQ_STOPS_DMA

DMA paused by DREQ state (RO)

END

End flag (W1C)

ERROR

DMA error (RO)

INT

Interrupt status (W1C)

OUTSTANDING_TRANSACTIONS

Indicates that there are outstanding AXI transfers, either outstanding read data or outstanding write responses (RO)

PANIC_QOS

AXI panic QoS level

QOS

AXI QoS Level

RD_PAUSED

DMA read paused state (RO)

WAITING_FOR_OUTSTANDING_WRITES

The DMA4 is waiting for all the write response to be returned (RO)

WAIT_FOR_OUTSTANDING_WRITES

Wait for outstanding writes

WR_PAUSED

DMA write paused state (RO)

Structs

Register

Constants

ABORT

ACTIVE

DISDEBUG

DMA_BUSY

DREQ

DREQ_STOPS_DMA

END

ERROR

INT

OUTSTANDING_TRANSACTIONS

PANIC_QOS

QOS

RD_PAUSED

WAITING_FOR_OUTSTANDING_WRITES

WAIT_FOR_OUTSTANDING_WRITES

WR_PAUSED